backend-tools/pyproject.toml
[project]
name = "backend-tools"
version = "0.1.0"
description = "API Tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "pydantic-settings>=2.13.0",
    "azure-devops>=7.1.0b4",
]
